pyproject.toml
[project]
name = "chat-seguran-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=41.0.0",
    "pynacl>=1.6.0",
    "websockets>=11.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.49.3",
    "pytest>=9.0.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["client", "server"]
